Security/03_injection/sql-injection-lab/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
bcrypt 
pydantic
pytest
httpx
pytest-asyncio
aiosqlite
asgi-lifespan